fibHandlerの分割
